sink 방법 didSet으로 변경
